claude-interaction/requirements.txt
anthropic>=0.7.0
pyautogui>=0.9.54
python-dotenv>=1.0.0